Dash-To-Dock hover feature added as a preference option
DashToDock hover .. It shows the workspaces dock when hovering over the dash-to-dock extension (if you've got it installed).
Extremely useful in cases where your workspaces dock is hidden and you want to open a new app from the dash. Rather than going into overview just to see your workspaces, hover over the dash-to-dock extension. The workspaces dock will show to the right. Use the dash-to-dock scroll to go to the appropriate workspace.