Thumbnail window images/clones are not properly representing maximized or fullscreen app windows that use Client Side Decorations.
To Duplicate:
Open gedit text editor (prime example of CSD app). Move it around on the desktop and you'll notice that the thumbnail window images stay in sync (that's good).

Now, move gedit to the center of the desktop and maximize it. Press <super> to enter activities overview (or just mouse over to the dock) and notice how the thumbnail window image is positioned slightly to the top-left leaving the workspace background to be seen on the right-bottom edge of the thumbnail. Now, unmaximize gedit and notice how the thumbnail window image is in a different position than the actual window on the desktop.

Move gedit to the center of the desktop again. Hit F11 to enter fullscreen mode. Hit <super> key to enter the activities overview and see how the thumbnail window image is positioned to the right leaving the workspace background to show on the left side of the thumbnail.