Fix bug where dock leaves gap near edge of display returning from lock screen.
DETAILS:
The issue only happens when dash-to-dock is set to fixed position and workspaces-to-dock is 
set to autohide on with intellihide off. What causes the issue is the layout manager updating 
the work area while the dock is animating.

OTHER NOTES:
A second issue observed while testing was that  workspaces-to-dock failed to update the 
thumbnails size when set to fixed position and returning from lock screen.

A third issue observed while testing was that workspaces-to-dock failed to update the 
thumbnails size when dash-to-dock was set to fixed position.

All issues are dealt with in this commit.